Moved Player to Chess module
